requirements.txt
PyMuPDF>=1.23.0
pdf2image>=1.16.3
easyocr>=1.7.1
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
openpyxl>=3.1.2
pdfplumber>=0.11.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
